+++ src/features/ModelPlan/ModelToOperations/_components/ActionsMenu/index.tsx
@@ -1,18 +1,13 @@
-import React, { useContext, useEffect, useMemo, useRef, useState } from 'react';
-import { useTranslation } from 'react-i18next';
-import { useHistory } from 'react-router-dom';
+import React, { useContext, useEffect, useRef, useState } from 'react';
 import { Button, Menu } from '@trussworks/react-uswds';
 import classNames from 'classnames';
 import i18next from 'i18next';
 
-import Modal from 'components/Modal';
-import PageHeading from 'components/PageHeading';
-import Sidepanel from 'components/Sidepanel';
+import { EditMTOMilestoneContext } from 'contexts/EditMTOMilestoneContext';
 import { MTOModalContext } from 'contexts/MTOModalContext';
 import useCheckResponsiveScreen from 'hooks/useCheckMobile';
 import useMessage from 'hooks/useMessage';
 
-import EditMilestoneForm from '../EditMilestoneForm';
 import { MTORowType } from '../MatrixTable/columns';
 
 const ActionMenu = ({
@@ -32,38 +27,13 @@ const ActionMenu = ({
   subCategoryID?: string;
   name?: string;
 }) => {
-  const { t: modelToOperationsMiscT } = useTranslation('modelToOperationsMisc');
-
   const [isMenuOpen, setIsMenuOpen] = useState<boolean>(false);
 
   const { clearMessage } = useMessage();
 
-  const history = useHistory();
-
   const { setMTOModalOpen, setMTOModalState } = useContext(MTOModalContext);
 
-  const params = useMemo(
-    () => new URLSearchParams(history.location.search),
-    [history]
-  );
-
-  const milestoneParam = params.get('edit-milestone');
-
-  const [isModalOpen, setIsModalOpen] = useState(
-    milestoneParam === milestoneID
-  );
-
-  const submitted = useRef<boolean>(false);
-
-  const [isDirty, setIsDirty] = useState<boolean>(false);
-
-  const [leavePage, setLeavePage] = useState<boolean>(false);
-
-  useEffect(() => {
-    if (milestoneParam === milestoneID) {
-      setIsModalOpen(true);
-    }
-  }, [milestoneParam, milestoneID, setIsModalOpen]);
+  const { openEditMilestoneModal } = useContext(EditMTOMilestoneContext);
 
   const isTablet = useCheckResponsiveScreen('tablet', 'smaller');
 
@@ -252,91 +222,14 @@ const ActionMenu = ({
       </div>
     );
 
-  const closeModal = () => {
-    if (isDirty && !submitted.current) {
-      setLeavePage(true);
-    } else if (!isDirty || submitted.current) {
-      params.delete('edit-milestone');
-      params.delete('select-solutions');
-      history.push({ search: params.toString() });
-      setLeavePage(false);
-      setIsModalOpen(false);
-      submitted.current = false;
-    }
-  };
-
   return (
     <div style={{ textAlign: 'right' }}>
-      <Sidepanel
-        isOpen={isModalOpen}
-        closeModal={closeModal}
-        ariaLabel={i18next.t(
-          'modelToOperationsMisc:milestoneLibrary.aboutThisMilestone'
-        )}
-        testid="edit-milestone-sidepanel"
-        modalHeading={i18next.t(
-          'modelToOperationsMisc:milestoneLibrary.aboutThisMilestone'
-        )}
-        noScrollable
-      >
-        <EditMilestoneForm
-          closeModal={closeModal}
-          setIsDirty={setIsDirty}
-          submitted={submitted}
-        />
-      </Sidepanel>
-
-      <Modal
-        isOpen={leavePage && !submitted.current}
-        closeModal={() => setLeavePage(false)}
-        className="confirmation-modal"
-      >
-        <PageHeading
-          headingLevel="h3"
-          className="margin-top-neg-2 margin-bottom-1"
-        >
-          {modelToOperationsMiscT('modal.editMilestone.leaveConfim.heading')}
-        </PageHeading>
-
-        <p className="margin-top-2 margin-bottom-3">
-          {modelToOperationsMiscT(
-            'modal.editMilestone.leaveConfim.description'
-          )}
-        </p>
-
-        <Button
-          type="button"
-          className="margin-right-4 bg-error"
-          onClick={() => {
-            params.delete('edit-milestone');
-            history.replace({ search: params.toString() });
-            setIsModalOpen(false);
-            setIsDirty(false);
-            setLeavePage(false);
-          }}
-        >
-          {modelToOperationsMiscT('modal.editMilestone.leaveConfim.confirm')}
-        </Button>
-
-        <Button
-          type="button"
-          unstyled
-          onClick={() => {
-            setLeavePage(false);
-          }}
-        >
-          {modelToOperationsMiscT('modal.editMilestone.leaveConfim.dontLeave')}
-        </Button>
-      </Modal>
-
       <Button
         type="button"
         unstyled
         className="margin-right-2"
         onClick={() => {
-          params.set('edit-milestone', milestoneID);
-          history.push({ search: params.toString() });
-          setIsModalOpen(true);
+          openEditMilestoneModal(milestoneID);
         }}
       >
         {i18next.t('modelToOperationsMisc:table.editDetails')}
